backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-core
langchain_community
langchain-groq
groq
python-dotenv